Don't switch to tmux session if popup selector is closed
Your fzf session jumper is awesome and I've used it thousands of
times. However, it had the annoying behaviour of switching to a
session even if the fzf popup is closed (with Esc or Ctrl+C).

This commit changes that by only executing `tmux switch-client` on
'enter', meaning that a cancel will simply quit the fzf popup and
leave you in the current session.